move mdformat plugin to its own file
