+++ scripts/_utils.py
@@ -2,15 +2,6 @@ from __future__ import annotations
 
 import re
 from pathlib import Path
-from typing import TYPE_CHECKING
-
-import mdformat
-
-if TYPE_CHECKING:
-    import argparse
-
-    from markdown_it import MarkdownIt
-    from mdformat.renderer import RenderContext, RenderTreeNode
 
 ROOT_DIR = Path(__file__).resolve().parent.parent
 
@@ -33,28 +24,3 @@ def snake_case(name: str) -> str:
 
 def get_indent(line: str) -> str:
     return re.match(r"^\s*", line).group()  # type: ignore[union-attr]
-
-
-class NoEscapeTextPlugin:
-    """Overrides the default text formatting behavior of mdformat."""
-
-    def __init__(self: NoEscapeTextPlugin) -> None:
-        self.POSTPROCESSORS = {"text": NoEscapeTextPlugin.text}
-        self.RENDERERS = {}
-
-    @staticmethod
-    def add_cli_options(parser: argparse.ArgumentParser) -> None:
-        pass
-
-    @staticmethod
-    def update_mdit(mdit: MarkdownIt) -> None:
-        pass
-
-    @staticmethod
-    def text(_text: str, node: RenderTreeNode, _context: RenderContext) -> str:
-        return node.content
-
-
-def add_no_escape_text_plugin() -> None:
-    """Add NoEscapeTextPlugin to the list of mdformat extensions."""
-    mdformat.plugins.PARSER_EXTENSIONS["no-escape-text"] = NoEscapeTextPlugin()
